backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP client for external APIs
httpx==0.25.2
requests==2.31.0

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS middleware
fastapi-cors==0.0.6

# OAuth2 and JWT
authlib==1.2.1
itsdangerous==2.1.2